// client/lexer/types.ts
export type Keyword =
  | "program"
  | "procedure"
  | "function"
  | "class"
  | "def"
  | "begin"
  | "end"
  | "endproc"
  | "if"
  | "then"
  | "else"
  | "elif"
  | "endif"
  | "for"
  | "to"
  | "downto"
  | "step"
  | "in"
  | "next"
  | "while"
  | "endwhile"
  | "do"
  | "repeat"
  | "until"
  | "pass"
  | "return"
  | "const"
  | "final"
  | "var"
  | "dim"
  | "local"
  | "private"
  | "global"
  | "nonlocal"
  | "array"
  | "of"
  | "void"
  | "boolean"
  | "character"
  | "integer"
  | "string";

export type Type = "boolint" | "boolean" | "integer" | "character" | "string";

export type Operator =
  | "neg"
  | "not"
  | "plus"
  | "subt"
  | "mult"
  | "divr"
  | "div"
  | "mod"
  | "and"
  | "or"
  | "xor"
  | "andl"
  | "orl"
  | "scat"
  | "eqal"
  | "noeq"
  | "less"
  | "more"
  | "lseq"
  | "mreq"
  | "seql"
  | "sneq"
  | "sles"
  | "smor"
  | "sleq"
  | "smeq";

export type Delimiter = "(" | ")" | "{" | "}" | "[" | "]" | "," | ":" | ";" | "." | ".." | "->";